+++ src/main/java/racing/controller/RacingController.java
@@ -1,14 +1,15 @@
 package racing.controller;
 
 import java.util.StringTokenizer;
+import racing.service.RacingService;
 import racing.service.RacingServiceImpl;
 
 public class RacingController {
 
-  private final RacingServiceImpl racingServiceImpl;
+  private final RacingService racingService;
 
-  public RacingController(RacingServiceImpl racingServiceImpl) {
-    this.racingServiceImpl = racingServiceImpl;
+  public RacingController(RacingService racingService) {
+    this.racingService = racingService;
   }
 
   public String startRace(StringTokenizer requestRegisterCars, int round) {
@@ -18,19 +19,19 @@ public class RacingController {
   }
 
   private void register(StringTokenizer requestRegisterCars) {
-    racingServiceImpl.register(requestRegisterCars);
+    racingService.register(requestRegisterCars);
   }
 
   private void start(int racingRoundCount) {
-    racingServiceImpl.start(racingRoundCount);
+    racingService.start(racingRoundCount);
   }
 
   private String getRacingProcess() {
-    return racingServiceImpl.getRacingProcess();
+    return racingService.getRacingProcess();
   }
 
   public String getWinner() {
-    return racingServiceImpl.getWinner();
+    return racingService.getWinner();
   }
